New algorithm modules created.
Algorithms separated to different modules.
